requirements.txt
arrow==1.2.2
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
croniter==1.3.5
Deprecated==1.2.13
Django==4.1
django-model-utils==4.2.0
django-rq==2.5.1
django-utils-six==2.0
djangorestframework==3.13.1
drf-yasg==1.21.3
Flask==2.2.2
idna==3.3
importlib-metadata==4.12.0
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
rq==1.11.0
rq-scheduler==0.11.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.11
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.8.1
